Déduplication de code
Avec la création d'une superclasse SokobanMovable, tout ce qui bouge
héritera de ça maintenant.
